+++ src/shell/commands/executable.rs
@@ -57,6 +57,8 @@ impl ShellCommand for ExecutableCommand {
         }
       };
 
+      context.state.track_child_process(&child);
+
       // avoid deadlock since this is holding onto the pipes
       drop(sub_command);
